+++ source/src/UStat.cpp
@@ -40,15 +40,18 @@
 namespace 
 {
   // ==========================================================================
-  double getDistance ( const RooArgSet* x , 
-                       const RooArgSet* y )
+  double getDistance
+  ( const RooArgSet* x , 
+    const RooArgSet* y )
   {
     if ( 0 == x || 0 == y ) { return -1 ; } // RETURN 
     //
     double result = 0.;
     //
-    Ostap::Utils::Iterator xIter ( *x ) ;
-    Ostap::Utils::Iterator yIter ( *y ) ;
+#if ROOT_VERSION_CODE < ROOT_VERSION(6,18,0)
+    //
+    Ostap::Utils::Iterator xIter ( *x ) ; // only for ROOT < 6.18 
+    Ostap::Utils::Iterator yIter ( *y ) ; // only for ROOT < 6.18 
     //
     RooRealVar* xVar = 0 ;
     RooRealVar* yVar = 0 ;
@@ -58,6 +61,21 @@ namespace
       const double val  = xVar->getVal() - yVar->getVal() ;
       result           += val*val ;
     }
+#else 
+    //
+    RooArgSet::const_iterator xIter = x -> begin () ;
+    RooArgSet::const_iterator yIter = y -> begin () ;
+    RooArgSet::const_iterator xEnd  = x -> end   () ;
+    RooArgSet::const_iterator yEnd  = y -> begin () ;    
+    for ( ; xIter != xEnd && yIter != yEnd ; ++xIter, ++yIter ) 
+    {
+      const RooRealVar* xVar  = static_cast<const RooRealVar*> ( *xIter ) ;
+      const RooRealVar* yVar  = static_cast<const RooRealVar*> ( *yIter ) ;
+      const double      val   = xVar->getVal() - yVar->getVal() ;
+      result                 += val*val ;
+    }
+    //
+#endif    
     //
     return std::sqrt ( result ) ;
   }
@@ -120,10 +138,23 @@ Ostap::StatusCode Ostap::UStat::calculate
     { return Ostap::StatusCode ( InvalidItem2 ) ; }             // RETURN 
     //
     // 2.Evaluate PDF 
-    Ostap::Utils::Iterator iter  ( *args ) ;
+    //
+#if ROOT_VERSION_CODE < ROOT_VERSION(6,18,0)
+    //
+    Ostap::Utils::Iterator iter  ( *args ) ; // only for ROOT < 6.18 
     RooRealVar * var = 0 ;
     while ( (var = (RooRealVar*)iter->Next() ) ) 
     { var->setVal( event_i->getRealValue( var->GetName() ) ); }
+    //
+#else 
+    //
+    for ( auto* vv : *args ) 
+    {
+      RooRealVar* var = static_cast<RooRealVar*> ( vv ) ;
+      var -> setVal ( event_i -> getRealValue ( var -> GetName () ) ) ;
+    }
+    //
+#endif
     //
     const double pdfValue = pdf . getVal( args ) ;
     //